requirements.txt
iridi
discord.py-self
